:recycle: refactor: added unmanaged versions of the int. group and list
